separated sidebar to own component
